fix getscom_abs and _getscom code
Instead of reading the scom registers, and storing the results
to a memory address specified by o_data, the two routines were
loading data into a vector register, without storing it to the
o_data pointer, basically losing it.

There is no memory to memory PPC instruction, so the minimum number
of instructions to execute the code is two: a vector load and a vector store.

- Had to modify the ppe42_scom.h to make the _m_data parameter to the
PPE_LVD and PPE_LVDX macros an output only. Otherwise, the compiler
initializes the temp variable, adding two (useless) instructions.

- make the PPE_LVDX macro have the same flavor as PPE_LVD by eliminating
the pointer to m_data from both the _getscom function side and the
macro implementation side.

Change-Id: I67a9eef16ed1e0f95f02ff8ba0336558cc8b23e3
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/20716
Tested-by: Jenkins Server
Reviewed-by: Claus Michael Olsen <cmolsen@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>