p9_ringId.H: resolve duplicate copies of file
As agreed with Prasad, the copy in chips/p9/utils is being abandoned,
while the copy in chips/p9/utils/imageProcs is maintained.

Both copies have diverged considerably.
It looks like the removed copy has never been updated.
So I am assuming that there is no code inside the removed copy
that needs to be kept and copied over into the retained copy.

Change-Id: Ie8b95842c07a44ea273764fad3186db279dc1a3d
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/24802
Tested-by: Jenkins Server
Tested-by: PPE CI
Reviewed-by: Claus M. Olsen <cmolsen@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: GIRISANKAR PAULRAJ <gpaulraj@in.ibm.com>
Tested-by: Hostboot CI
Reviewed-by: Martin Peschke <mpeschke@de.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/24803
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>