SBE Image Build Flow: no 2kB .overrides placeholder anylonger
We used to reserve 2kB inside the SBE Image for .overrides,
which was assumed to be replaced or overwritten with real
override rings. The requirements have been relaxed, the
placeholder is being removed. This will simplify
the placement of .override content from a tools perspective,
making the placement similar to what we will do as to .hbbl:
we simply call p9_xip_append().

Change-Id: I9b8268cc63d568baf81f889c97aa95c2afdbcdc1
RTC: 157404
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/26702
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: Claus M. Olsen <cmolsen@us.ibm.com>
Reviewed-by: Martin Peschke <mpeschke@de.ibm.com>