p9_putRingUtils: clean up duplicated code
p9_putRingUtils had a copy of the RS4 container data structure
from p9_scan_compression. Code duplication is bad. In this particular
case we are going to make changes to the RS4 container data structure
in a follow-up commit. Hence it is the right point in time to clean up
that code duplication.

Change-Id: Ia047ced59b4c924635b0c2f428b15a652db645d7
RTC: 158078
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/27695
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Martin Peschke <mpeschke@de.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/27698
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>