Fix some trace problems
This change is in preparation for future, better trace handling.
Bit fields don't have a size_type - so they need a cast to int or uint in a
trace statement.
Pointers are always 4 bytes in PPE.
Pointers cannot be converted to a uint32_t (%x) without an explicit cast.
The printf/trace format %x is 4 bytes, even %016x is 4 bytes. What is
needed is %llx or %016llx otherise a uint64_t will get truncated.

Change-Id: I0b55a355f3211a7f6913e846834486b4808f50c1
RTC: 161851
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31189
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>