p9_tor: fix endianess conversion
The p9_tor code used htobe* in many places where it should have used
be*toh, because the code reads values from the big-endian TOR image
to host memory. This misuse should not show during execution, because
these functions are equivalent. But the code was misleading and raising
the wrong questions (How can we ever use values converted from host
endian to big-endian in calculations and as sizes or offsets
in memcpy operation?)

Change-Id: I2117a22e3e0765dff44279e2bb291e13811b2f15
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/33317
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Martin Peschke <mpeschke@de.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/33320
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>