p9_fastarray: Fix scanning after dump, make dump repeatable
Change the cleanup code to let the ABIST run to completion and
deassert ABIST mode, so that the ABIST engines end up in a clean
state, enabling repeated dumps of the same core.

Also leaving ABIST mode asserted caused scan access to ec_abst
to fail -- this is fixed now too.

Plus add an option to the wrapper to force non-sparse scanning.

Change-Id: I3eac75878e589e13a506eb67031778964e6770ab
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36029
Reviewed-by: NAGENDRA K. GURRAM <nagendra.g@in.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36045
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>