Update proc_select_ex to remove target from old MC group
proc_select_ex sets up some multicast groups for Cores EXs and Eqs.
The code was not written to be called multiple times because
it was not removing the target from its old group when it was added
to a new group. Before we were blindly removing every target from
group 7 (which has nothing in it), now we read the prev group and
remove the target from the prev group

Change-Id: I8d90413855e862c93787cfa4700e1adf6ad9ee4a
Original-Change-Id: Iea339cb778cd50a2712f59ee41c243dcbae95d0c
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/37949
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: SRINIVAS V. POLISETTY <srinivan@in.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>