p9_fastarray: Add support for multiple chips, Nimbus DD2 support
Refactor engine code to switch from a fixed array and ring table to
a table per chip and EC level. Change a lot of loop constructs to
C++11-y foreach loops in the process.

Add a function to choose the right tables based on chip type and EC.

Also add Nimbus DD2 array and ring definitions.

Change-Id: I9d9ffeacc8bfe7f43a7afc7b73d5b831fb8db354
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40078
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: James N. Klazynski <jklazyns@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40130
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>