p9_sbe_chiplet_reset: Revert NX_1 hang pulse back to 34s
The hang pulse setting of 68s is not supported by the PCB slave,
and NX can deal with a 34s hang pulse too, so revert back to 34s.

We could change the chiplet's base divider and adapt all other hang
pulse settings, but that would be a huge code ripup as it breaks
uniformity among chiplets.

Change-Id: I17ae92e58d713d54256083f43eabd9ce4be7167f
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39795
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jenny Huynh <jhuynh@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39797
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>