parseAttributeInfo: Support inverted EC features
There are several EC feature attributes where matching the chip
specification actually indicates the _absence_ of a feature.

So far we had to either match on every other chip, adding new chips
as projects are added, or define the attribute as a negative attribute
and invert its value in procedure code (sometimes without indicating
in the attribute name that it's a negative attribute). Both approaches
are prone to causing mistakes down the road.

To fix this, introduce a new <falseIfMatch/> tag that indicates that
an EC feature is absent if the chip spec is matched, so the procedure
code will not have to remember to invert the attribute value.

Change-Id: I33dcf127b8f81729c9638d6648faf2c5adbcb470
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42608
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Santosh S. Puranik <santosh.puranik@in.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>