Add pm_check_quiesce function to list of sbe_check_quiesce function
In addition to powman suspend we need to make sure that special wakeups
are in a safe state when we do an MPIPL. We are adding in that cleanup
to sbe_check_quiesce. In this commit I also added in some other pm
cleanup that will be done later by p9_powman_suspend.

Change-Id: I8a24bec460c5a8185365ff8e749f92f938555de7
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/43213
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Brian T. Vanderpool <vanderp@us.ibm.com>
Reviewed-by: Michael S. Floyd <mfloyd@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/43218
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>