Fix bug in npu quiesce sequence
There were a couple issues w/ the sequence of steps we were doing
in order to quiesce the npu. The first problem was that we were
doing an indirect scom incorrectly and instead of the actual value
we were getting all 1's back. After fixing that we realized that
the NTLs were not making it to the reset state, this was because
the brick was never enabled by PHYP. A check was added to not
attempt to put NTLs in reset state(step 1) if brick isnt enabled

CQ: SW397252
Change-Id: I94899c73438d8bbdaf777f8aab5640992ae74db6
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44707
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: James F. Mikos <mikos@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44710
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>