Move clearing of CPMMR PPM WRITE DISABLE so it called on all func cores
This HWP loops over functional cores while selecting the master. The
loop exits after it detects all of the master exs in the master quad.
The logic that clears the CPMMR WRITE DISABLE was only getting called
on the master core in this loop. I added another loop over cores
which calls this logic so this bit will be cleared on all cores.

Change-Id: Icc5fa3285f9d6710eec665dcb6cc7ad16d5bbf88
Original-Change-Id: I023e18db961a23cd0d1181ab558f61c8df8d4ec3
CQ: SW399710
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44975
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: YUE DU <daviddu@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>