Ensure SGPE is disabled and ensure writes are enabled during pm_suspend
Previously if the PGPE was halted the pm_suspend HWP would not do anything.
This caused and issue if an MPIPL was triggered in a window where the SGPE
was running but the PGPE was not. This commit adds a check to see if the
SGPE isnt halted, and the PGPE is, if that is true then the SGPE gets
halted via the XIXCR. In addition to this change we have were seeing
issues where if an MPIPL was triggered in the window where SGPE isnt halted
and the PGPE is, then the PCB_WRITE_DISABLE bit was never getting cleared
which caused problems when we tried to power down the quads.

Change-Id: If2c67ce1153a949527b8e9f2cfcba1e0fe04c398
CQ: SW398122
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/45444
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Brian T. Vanderpool <vanderp@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/45448
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>