Remove functionality from tp enable ridi and move it to nest enabled ridi
Previously, we were enabled TP RIDI before we did a bunch of scans which
touch LPC logic. This was causing LPC logic to get messed up because
once TP ridi is enabled LPC traffic is flowing. To get around this,
we are moving the enablement of the TP RIDI to after the scans of the
LPC logic so that the LPC logic wont get messed with

Change-Id: I6244fdf1314a21d9c76519bde3905287c7870b26
CQ: SW396004
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/46941
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/46966
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>