Increase cache data timeout values
The PCIe timeout have increased to 66-100ms, so we need to double
the NCU master timeouts to be above these timeouts. This has a
chain effect causing the L2/L3 master timeouts to increase which
causes the tlbie snooper to increase which causes the tlbie master
to increase.

This would also ususally cause the core timeout to increas, but
the core is already at around 13 seconds, so there is headroom
there.

Change-Id: I5930076151267a9bfa66e24edef0985c165db0b7
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50582
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: LUKE MURRAY <murrayl@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50602
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>