p9_sbe_check_quiesce -- restore call to p9_int_scrub_caches
52512 removed code related to the DD1 SW based INT reset sequence, to leave
only the HW based reset in production code for DD2 and beyond.  It also
erroneously removed the call to/code for p9_int_scrub_caches.

This commit restores the subroutine, and invokes it prior to the HW
quiesce/sync reset into order to scrub/flush the EQC, VPC, IVC, and SBC caches.

Change-Id: I051117e3a18c55aea7267e53eea1652f0cff9790
CQ: SW431898
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/62227
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Jenny Huynh <jhuynh@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/62243
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>