Clear INT_CQ related firs after completing sync_reset in MPIPL
In the SBE steps of the MPIPL, the SBE calls a HWP called
p9_sbe_check_quiesce. This function ensures that traffic on the
powerbus is stopped prior to cycling the master core on and off.
During this HWP a sync_reset is performed on the INT component.
After this reset we have been told to clear out all of the related'
INT_CQ firs. This commit adds in a new function which is called
immediatly after sync reset and will clear all relevent scoms.

Change-Id: Ia99a2f2d3f855823472f81b32baf44d25d7c4cad
CQ:SW448121
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68020
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68023
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>