Leave scratch valid bits alone to allow HB to query
Current behavior  of SBE is to set all the scratch 8
valid bits for all of the scratch registers, always.
This behavior makes it impossible for Hostboot to determine
if there was an actual override.   Since the SBE always ensures
valid settings in the scratch regs, leave the valid bits in
original state so Hostboot can make its own determination.

Change-Id: Iade385b289cb2300a4b7ed745981c4cac95a8b9c
CQ: SW451891
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69062
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69090
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>