Update i2c driver for OCMB chip
After doing some testing with the simics Axone model we found that
we had some errors in the protocols used to do reg reads and writes.
We had to put extra dummy information when we latch and address to
read and also we were not reading enough bytes when reading the
status. This commit addresses both of these issues.

Change-Id: I78f1626047700c5a5b7f2e8b3934a9d2f5220dfb
Original-Change-Id: I15ce40ef28d6c6c2303fe600986c8872b0829ad0
RTC: 196806
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69575
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80979
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>
Tested-by: RAJA DAS <rajadas2@in.ibm.com>