apply HW423589 option1 (MCD disable) workaround for p9n DD2.1
Current FW uses option2 (alternate MCD configuration) to workaround
HW423589 for p9n DD2.1 hardware, which is desired for OP releases
supporting GPUs.

This commit adjusts the DD2.1 solution to option1 instead,
and can be cherry-picked into the enterprise release branches to
remove the per-socket memory resrictions which come as a side effect
of the current option2 setup.

Change-Id: I24ecf918a81964a3df6b1e3e18e60375d7646b45
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69718
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Kevin F. Reick <reick@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69845
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>