p9_sbe_npll_setup: Enable spread spectrum before starting PLL
There's a synchronizer missing on the sscgen input of the filter
PLLs, thus any changes to that input must be made before the PLL
is taken out of reset, or the spread generator inside the PLL might
start up in a skewed state, resulting in a broken spread profile.

Thus, change the procedure to enable spread before taking the PLL
out of reset, and also make sure we never try to synchronously
enable spread.

Change-Id: Ie0e545a6baf492d394a7d63d99ba5b83a0c46423
CQ: SW457204
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71886
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Kevin F. Reick <reick@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71912
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>