Update cmd/rsp endian handling in exp_inband
Data written/read from the command/response/data
buffers is done so in transaction sizes of 4 or 8
bytes.  The data in those transactions will be
little endian byte order.  The byte order of those
transactions must be corrected to represent the
data structures stored in the buffers correctly
for extraction of field data.  The fields themselves
will be big endian byte ordering.

This patch also adds two attributes:

ATTR_MSS_OCMB_EXP_STRUCT_MMIO_ENDIAN_CTRL -
Controls whether or not bytes read from and
written two the buffer are swapped.  The default
is to swap the byte order.

ATTR_MSS_OCMB_EXP_STRUCT_ENDIAN -
Controls whether or not the bytes of buffer
structure fields are little endian or not.
The default is big endian.

Change-Id: I0863607bac383d586d10025a6c490cae0e2ed38f
Original-Change-Id: I734d25dea2a3e4a864a2d35df2576676cad27cfe
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72314
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamin Gass <bgass@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>