Renaming and relocating core image infrastructure files (new)
Note that this commit only renames and relocates key codes and gets rid
of all Centaur support in CII codes. It does not introduce any p10 meta-
data, but merely prepares the build environment and some key CII codes
for upcoming p10 ring and image support. The commit also includes
copying and renaming p9_xip_customize to the p10 dir as p10_ipl_customize
and fixing up, and copying, the relevant makefiles and XML files to ensure
a successful build of the p10_ipl_customize HWP. Even though we do not
need the HWP for anything at this point, moving the HWP into the p10
dir causes me to have to deal with the build environment to gain
experience how to handle this for ALL of our CII files and which is a
required learning curve on my part in anticipation of all the work
that needs to be done for the near term p10 VBU ring support.

Detailed mods:
- Copied all *.mk files for asis p9 codes into corresponding p10 dirs
  and added relevant p9 source include paths
- Copied and/or renamed the following APIs:
  - ./chips/p9/.../p9_tor.* ->
    ./chips/p10/.../p10_tor.*
  - ./chips/p9/.../p9_dd_container.c|h|mk ->
    ./chips/p10/.../p10_ddco.C|H|mk
  - ./chips/p9/xip/p9_xip_image.C|h|mk ->
    ./chips/p10/utils/imageProcs/p10_ipl_image.C|H|mk
- Copied and renamed the following tools:
  - ./chips/p9/xip/p9_xip_tool.C|mk ->
    ./tools/imageProcs/ipl_image_tool.C|mk
  - ./chips/p9/.../p9_dd_container_tool.c|mk ->
    ./tools/imageProcs/ddco_tool.C|mk
  - ./ring_apply/p9_ring_apply.C|mk ->
    ./tools/imageProcs/ring_apply.C|mk
  - p9_ipl_build.* ->
    ipl_build.*
- Copied and renamed the following HWPs:
  - ./chips/p9/.../customize/p9_xip_* ->
    ./chips/p10/.../customize/p10_ipl_*
  - ./chips/p9/.../customize/wrapper/p9_xip_* ->
    ./chips/p10/.../customize/wrapper/p10_ipl_*
  - ./chips/p9/.../xml/*_info/p9_xip_* ->
    ./chips/p10/.../xml/*_info/p10_ipl_*
- Copied all, and renamed some, of the Cronus wrappers for TOR,
  RS4 and IPL (formerly XIP) image API access to ->
  ./chips/p10/utils/imageProcs/wrapper/*
- Disabled all customize image updates wrt attributes using the
  PPD #ifdef CHIP_GEN_P9
- Updated p10_ipl_customize to ALWAYS assume there is GPTR rings
  in Mvpd and there is GPTR overlays in the HW image.
- Eliminated of all Centaur IPL image and API support in CII
  codes

Notes:
- I've had to remove p9_ipl_build.* since I had to make changes
  to it and thus had to copy and rename it (I chose ipl_build.*
  since it's in the common tools dir). However, this causes
  p9_ipl_build.mk file to be picked up by ekb build and which
  can't be built since some of the p9 libs are no longer built
  if they have been renamed, e.g. p9_dd_container, or modified
  for p10, e.g. p9_tor and p9_xip_image..
- There is a lot of p9 user codes, particularly HWPs, which
point to CII headers and libs, for example p9_hcode_image_build and
p9_io_x/obus_image_build, that need to be updated to pick up the
relocated or renamed headers and libs. I've decided for now to
leave these codes alone since they do not prevent the ekb build.
The respective owners, for example Chris Steffen and Prem,  have to
be alert to picking up the relocated or renamed CII codes though
when they get around to making p10 specific changes in their codes.

Change-Id: I68c65a1f2ced8a1947d711edcb76f5af5694c021
Original-Change-Id: I480cccac51d4e179f6b5798112a66b4641045d7d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72538
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Reviewed-by: Kahn C. Evans <kahnevan@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>