Update p9_l2_flush to check if purge is busy on anything prior to flush.
The previous check was only checking once if CMD_REG_BUSY is off.  This
will now check if CMD_PRGSM_BUSY is off, indicating the Purge engine
is not busy with any requests at all.  It will also poll on busy
prior to issuing the flush.

If PURGE_CMD_ERR is on prior to starting flush a warning will be
printed via FAPI_DBG.

PURGE_CMD_ERR will be cleared prior to issuing the purge.

Change-Id: I120cc9a00d26da8cf2ca4ec6dd7d8f3006633b61
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72562
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: MURULIDHAR NATARAJU <murulidhar@in.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72582
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>