Updated OTPROM to update for clock divider setting if requested
Updated the OTPROM code to look for the SPI clock divider setting.
If the valid setting are found, then the same will be applied on
the SPI for the measurement seeprom.
The code expects the divider setting at 0x50039, at the MSB 4
nibbles. Clock divider bits are only 12bits in the SPI, so the top
nibble will be dropped.
The divider value will be saved in local register 0xc0002040

Change-Id: I67acbaa2e663fc55afd66822a23c28cfb0053b71
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/79429
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>