p10_ringId: Add inline function to query the name of a ring
There's another API function for this already (ringidGetRingName),
but it involves translation into a std::string and back, evaluating
return codes and is only available in Cronus so requires #ifdefs in
code that tries to use it.

Add a simpler variant for P10 HWP usage that's fixed to P10, returns
a const pointer into an existing table and an empty string on PPE
platforms.

Change-Id: I004f416506acd1be3d0b00c17e78a0d5223d50b1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81964
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>