Update ocmb i2c driver to send address in read command sequence
As part of the register read operation to a ocmb card the spec tells
us to perform 2 commands. The first command is a REGISTER_LATCH cmd
which includes the address of the register you want to read. The 2nd
cmd is a REGISTER_READ command which, according to the spec does not
require the address to be sent again. We found during bringup that
the gemini cards were expecting the address to be sent along with
the REGISTER_READ cmd as well as with the REGISTER_LATCH cmd. Cronus
was doing this in their code so they did not see this problem. As
these address bits will be ignored by explorer cards (per the spec)
it should be benign the include them with the REGISTER_READ cmd.

Change-Id: Ifcade18a290355432df43c0c3e58d6575f1f25b5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82249
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82257
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>