Fast Array dump, SBE part
This procedure dumps an entire core's arrays according to control data that
specifies which ABIST cycles x ring bits we actually care about, and streams
back the collected bits.

The data returned is raw ring data and must be post-processed by an offline
tool to yield the reconstructed array contents.

Change-Id: Ied6ee672b30d2b8814cbc5e7ff8a42076dd5e03e
Original-Change-Id: I02719b47ec77240f9e581b852371319b909ac0a8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85577
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-by: Christopher M Riedl <cmr@ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101652
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>