p10_sbe_dts_init -- multicast group updates
Cronus noted that this istep was missing from their istep support and added it
this week.  The current HWP code was last touched before some of the final
multicast group changes went in and is currently failing.

This commit simply changes the EQ group referenced to MCGROUP_ALL_EQ,
which is expected to be used by HWPs initailizing EQ resources in istep 3 prior
to p10_sbe_select_ex.

Change-Id: Ia33f97ba85c882f07e6598c16fe26a3aa59eb093
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85091
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Christopher M Riedl <cmr@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85092
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>